posixify day-4's Converter.cpp
So it can compile on Linux and Windows POSIX environments (read MSYS2). Since
this is a C++ file, I could write it in a portable way, using getchar() and
C++17's OS agnostic directory traversal facilities. But I thought of that after
the fact, and I'm not sure if anyone cares. Let's see what's in store for the
rest of the tutorial before coming back to this.